requirements.txt
anyio==3.6.1
click==8.1.3
fastapi==0.79.0
Flask==2.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.12.0
injector==0.20.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pydantic==1.9.2
pydotenv==0.0.7
pymongo==4.1.1
python-dotenv==0.20.0
PyYAML==6.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
Werkzeug==2.1.2
zipp==3.8.0
numpy